requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
gunicorn
psycopg2-binary
python-dotenv